requirements.txt
pandas
requests
python-dotenv
plotly
streamlit

